optimized inner loops of writer & reader threads
and bumped amount of words to send to 2E9.

It appears that inner cycle that's filling memory to send was to put
it midle under-optimized by gcc-4.7. I made it more straightforward
which fixed problem.

Also I fixed small problem that caused us to send a bit more data than
asked for.